+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_vcf/AnnotateVCFCommandLineParser.java
@@ -48,6 +48,8 @@ public class AnnotateVCFCommandLineParser extends JannovarAnnotationCommandLineP
 		result.writeVCFAnnotationStandardInfoFields = !cmd.hasOption("no-new-info-field");
 		result.escapeAnnField = !cmd.hasOption("no-escape-ann-field");
 		result.nt3PrimeShifting = !cmd.hasOption("no-3-prime-shifting");
+		if (cmd.hasOption("output-infix"))
+			result.outputInfix = cmd.getOptionValue("output-infix");
 
 		String args[] = cmd.getArgs(); // get remaining arguments
 		if (args.length < 3)
@@ -80,14 +82,16 @@ public class AnnotateVCFCommandLineParser extends JannovarAnnotationCommandLineP
 				"do not escape characters in INFO field \"ANN\" (default is on)"));
 		options.addOption(new Option(null, "no-3-prime-shifting", false,
 				"disable shifting of variants towards the 3' end of the transcript (default is on)"));
+		options.addOption(new Option(null, "output-infix", true,
+				"output infix to place before .vcf/.vcf.gz/.bcf in output file name (default is \".jv\")"));
 	}
 
 	private void printHelp() {
 		final String HEADER = new StringBuilder()
-		.append("Jannovar Command: annotate\n\n")
-		.append("Use this command to annotate a VCF file.\n\n")
-		.append("Usage: java -jar de.charite.compbio.jannovar.jar annotate [options] <database> [<IN.VCF>]+\n\n")
-		.toString();
+				.append("Jannovar Command: annotate\n\n")
+				.append("Use this command to annotate a VCF file.\n\n")
+				.append("Usage: java -jar de.charite.compbio.jannovar.jar annotate [options] <database> [<IN.VCF>]+\n\n")
+				.toString();
 		final String FOOTER = new StringBuilder().append(
 				"\n\nExample: java -jar de.charite.compbio.jannovar.jar annotate data/hg19_ucsc.ser IN.vcf\n\n")
 				.toString();