+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## develop
 
+### jannovar-cli
+
+* Improving output file generation, jannovar-cli now uses the same extension
+  as in the input and the infix is configurable instead of being fixed to
+  ".jv".
+* Default extension is ".vcf.gz" instead of ".vcf" now.
+
 ### jannovar-core
 
+* Fixing label for `FRAMESHIFT_VARIANT` in `VariantEffect`.
 * Moving CompatibilityCheckerException to package
   `...jannovar.pedigree.compatibilitychecker`
 * Fixing bug in transcript coordinate projection.